Moving app-deploy generation to build and dropping knative support
